requirements.txt
requests
bs4
lxml
dnslib
aiohttp
psycopg2-binary
tldextract
python-whois
flask
aiodns
pytz
reportlab
